Include python-apt pkg with initial Python install
This resolves the warning that this Ansible dependency is not
present in later task steps.

Note: It's been long enough since I worked with this playbook
that I no longer recall why I am duplicating the default
settings included in the `ansible-role-lxd-testenv` role for
Ubuntu systems. I have to assume I did so in an effort to
illustrate that the command could be overridden with
group-specific settings.

refs GH-48
refs atc0005/ansible-role-lxd-testenv#50